switched to using query params and other fixes
